Update sdk to support rhone upgrade

+++ packages/web3/src/contract/contract.ts
@@ -88,7 +88,7 @@ const crypto = new WebCrypto()
 export type FieldsSig = node.FieldsSig
 export type MapsSig = node.MapsSig
 export type EventSig = node.EventSig
-export type FunctionSig = node.FunctionSig
+export type FunctionSig = Omit<node.FunctionSig, 'isPublic' | 'usePreapprovedAssets' | 'useAssetsInContract'>
 export type Fields = NamedVals
 export type Arguments = NamedVals
 export type Constant = node.Constant
@@ -281,7 +281,18 @@ export class ProjectArtifact {
     return fsPromises.writeFile(filepath, content)
   }
 
-  needToReCompile(compilerOptions: node.CompilerOptions, sourceInfos: SourceInfo[], fullNodeVersion: string): boolean {
+  getChangedSources(sourceInfos: SourceInfo[]): SourceInfo[] {
+    const result: SourceInfo[] = []
+    for (const sourceInfo of sourceInfos) {
+      const info = this.infos.get(sourceInfo.name)
+      if (typeof info === 'undefined' || info.sourceCodeHash !== sourceInfo.sourceCodeHash) {
+        result.push(sourceInfo)
+      }
+    }
+    return result
+  }
+
+  needToReCompile(compilerOptions: node.CompilerOptions, fullNodeVersion: string): boolean {
     ProjectArtifact.checkCompilerOptionsParameter(compilerOptions)
     if (this.fullNodeVersion !== fullNodeVersion) {
       return true
@@ -293,16 +304,6 @@ export class ProjectArtifact {
     })
     if (!optionsMatched) {
       return true
-    }
-
-    if (sourceInfos.length !== this.infos.size) {
-      return true
-    }
-    for (const sourceInfo of sourceInfos) {
-      const info = this.infos.get(sourceInfo.name)
-      if (typeof info === 'undefined' || info.sourceCodeHash !== sourceInfo.sourceCodeHash) {
-        return true
-      }
     }
 
     return false
@@ -327,15 +328,22 @@ export class ProjectArtifact {
   }
 }
 
-function removeOldArtifacts(dir: string) {
+function removeOldArtifacts(dir: string, sourceFiles: SourceInfo[]) {
   const files = fs.readdirSync(dir)
   files.forEach((file) => {
     const filePath = path.join(dir, file)
     const stat = fs.statSync(filePath)
     if (stat.isDirectory()) {
-      removeOldArtifacts(filePath)
+      removeOldArtifacts(filePath, sourceFiles)
     } else if (filePath.endsWith('.ral.json') || filePath.endsWith('.ral')) {
-      fs.unlinkSync(filePath)
+      const filename = path.basename(filePath)
+      const artifactName = filename.slice(0, filename.indexOf('.'))
+      const sourceFile = sourceFiles.find(
+        (s) => s.name === artifactName && (s.type === SourceKind.Contract || s.type === SourceKind.Script)
+      )
+      if (sourceFile === undefined) {
+        fs.unlinkSync(filePath)
+      }
     }
   })
 
@@ -522,7 +530,11 @@ export class Project {
     return fsPromises.writeFile(filePath, JSON.stringify(structs, null, 2))
   }
 
-  private async saveArtifactsToFile(projectRootDir: string): Promise<void> {
+  private async saveArtifactsToFile(
+    projectRootDir: string,
+    skipSaveArtifacts: boolean,
+    changedSources: SourceInfo[]
+  ): Promise<void> {
     const artifactsRootDir = this.artifactsRootDir
     const saveToFile = async function (compiled: Compiled<Artifact>): Promise<void> {
       const artifactPath = compiled.sourceInfo.getArtifactPath(artifactsRootDir)
@@ -532,9 +544,32 @@ export class Project {
       }
       return fsPromises.writeFile(artifactPath, compiled.artifact.toString())
     }
-    this.contracts.forEach((contract) => saveToFile(contract))
-    this.scripts.forEach((script) => saveToFile(script))
-    this.saveStructsToFile()
+    for (const [_, contract] of this.contracts) {
+      if (!skipSaveArtifacts || changedSources.find((s) => s.name === contract.sourceInfo.name) !== undefined) {
+        await saveToFile(contract)
+      }
+    }
+    for (const [_, script] of this.scripts) {
+      await saveToFile(script)
+    }
+    await this.saveStructsToFile()
+    await this.saveProjectArtifact(projectRootDir, skipSaveArtifacts, changedSources)
+  }
+
+  private async saveProjectArtifact(projectRootDir: string, skipSaveArtifacts: boolean, changedSources: SourceInfo[]) {
+    if (skipSaveArtifacts) {
+      // we should not update the `codeHashDebug` if the `skipSaveArtifacts` is enabled
+      const prevProjectArtifact = await ProjectArtifact.from(projectRootDir)
+      if (prevProjectArtifact !== undefined) {
+        for (const [name, info] of this.projectArtifact.infos) {
+          if (changedSources.find((s) => s.name === name) === undefined) {
+            const prevInfo = prevProjectArtifact.infos.get(name)
+            info.bytecodeDebugPatch = prevInfo?.bytecodeDebugPatch ?? info.bytecodeDebugPatch
+            info.codeHashDebug = prevInfo?.codeHashDebug ?? info.codeHashDebug
+          }
+        }
+      }
+    }
     await this.projectArtifact.saveToFile(projectRootDir)
   }
 
@@ -588,7 +623,9 @@ export class Project {
     contractsRootDir: string,
     artifactsRootDir: string,
     errorOnWarnings: boolean,
-    compilerOptions: node.CompilerOptions
+    compilerOptions: node.CompilerOptions,
+    changedSources: SourceInfo[],
+    skipSaveArtifacts = false
   ): Promise<Project> {
     const removeDuplicates = sourceInfos.reduce((acc: SourceInfo[], sourceInfo: SourceInfo) => {
       if (acc.find((info) => info.sourceCodeHash === sourceInfo.sourceCodeHash) === undefined) {
@@ -640,20 +677,23 @@ export class Project {
       errorOnWarnings,
       projectArtifact
     )
-    await project.saveArtifactsToFile(projectRootDir)
+    await project.saveArtifactsToFile(projectRootDir, skipSaveArtifacts, changedSources)
     return project
   }
 
   private static async loadArtifacts(
     provider: NodeProvider,
     sourceInfos: SourceInfo[],
-    projectArtifact: ProjectArtifact,
     projectRootDir: string,
     contractsRootDir: string,
     artifactsRootDir: string,
     errorOnWarnings: boolean,
     compilerOptions: node.CompilerOptions
   ): Promise<Project> {
+    const projectArtifact = await ProjectArtifact.from(projectRootDir)
+    if (projectArtifact === undefined) {
+      throw Error('Failed to load project artifact')
+    }
     try {
       const contracts = new Map<string, Compiled<Contract>>()
       const scripts = new Map<string, Compiled<Script>>()
@@ -699,7 +739,8 @@ export class Project {
         contractsRootDir,
         artifactsRootDir,
         errorOnWarnings,
-        compilerOptions
+        compilerOptions,
+        sourceInfos
       )
     }
   }
@@ -822,19 +863,22 @@ export class Project {
     projectRootDir = '.',
     contractsRootDir = Project.DEFAULT_CONTRACTS_DIR,
     artifactsRootDir = Project.DEFAULT_ARTIFACTS_DIR,
-    defaultFullNodeVersion: string | undefined = undefined
+    defaultFullNodeVersion: string | undefined = undefined,
+    skipSaveArtifacts = false
   ): Promise<void> {
     const provider = getCurrentNodeProvider()
     const fullNodeVersion = defaultFullNodeVersion ?? (await provider.infos.getInfosVersion()).version
     const sourceFiles = await Project.loadSourceFiles(projectRootDir, contractsRootDir)
     const { errorOnWarnings, ...nodeCompilerOptions } = { ...DEFAULT_COMPILER_OPTIONS, ...compilerOptionsPartial }
     const projectArtifact = await ProjectArtifact.from(projectRootDir)
+    const changedSources = projectArtifact?.getChangedSources(sourceFiles) ?? sourceFiles
     if (
       projectArtifact === undefined ||
-      projectArtifact.needToReCompile(nodeCompilerOptions, sourceFiles, fullNodeVersion)
+      projectArtifact.needToReCompile(nodeCompilerOptions, fullNodeVersion) ||
+      changedSources.length > 0
     ) {
       if (fs.existsSync(artifactsRootDir)) {
-        removeOldArtifacts(artifactsRootDir)
+        removeOldArtifacts(artifactsRootDir, sourceFiles)
       }
       console.log(`Compiling contracts in folder "${contractsRootDir}"`)
       Project.currentProject = await Project.compile(
@@ -845,21 +889,21 @@ export class Project {
         contractsRootDir,
         artifactsRootDir,
         errorOnWarnings,
-        nodeCompilerOptions
-      )
-    } else {
-      console.log(`Contracts are compiled already. Loading them from folder "${artifactsRootDir}"`)
-      Project.currentProject = await Project.loadArtifacts(
-        provider,
-        sourceFiles,
-        projectArtifact,
-        projectRootDir,
-        contractsRootDir,
-        artifactsRootDir,
-        errorOnWarnings,
-        nodeCompilerOptions
+        nodeCompilerOptions,
+        changedSources,
+        skipSaveArtifacts
       )
     }
+    // we need to reload those contracts that did not regenerate bytecode
+    Project.currentProject = await Project.loadArtifacts(
+      provider,
+      sourceFiles,
+      projectRootDir,
+      contractsRootDir,
+      artifactsRootDir,
+      errorOnWarnings,
+      nodeCompilerOptions
+    )
   }
 }
 
@@ -876,24 +920,22 @@ export abstract class Artifact {
 
   abstract buildByteCodeToDeploy(initialFields: Fields, isDevnet: boolean): string
 
-  publicFunctions(): string[] {
-    return this.functions.filter((func) => func.isPublic).map((func) => func.name)
-  }
-
-  usingPreapprovedAssetsFunctions(): string[] {
-    return this.functions.filter((func) => func.usePreapprovedAssets).map((func) => func.name)
-  }
-
-  usingAssetsInContractFunctions(): string[] {
-    return this.functions.filter((func) => func.useAssetsInContract).map((func) => func.name)
-  }
-
   async isDevnet(signer: SignerProvider): Promise<boolean> {
     if (!signer.nodeProvider) {
       return false
     }
     const chainParams = await signer.nodeProvider.infos.getInfosChainParams()
     return isDevnet(chainParams.networkId)
+  }
+}
+
+function fromFunctionSig(sig: node.FunctionSig): FunctionSig {
+  return {
+    name: sig.name,
+    paramNames: sig.paramNames,
+    paramTypes: sig.paramTypes,
+    paramIsMutable: sig.paramIsMutable,
+    returnTypes: sig.returnTypes
   }
 }
 
@@ -911,6 +953,7 @@ export class Contract extends Artifact {
 
   readonly bytecodeDebug: string
   readonly codeHashDebug: string
+  readonly decodedMethods: Method[]
 
   constructor(
     version: string,
@@ -942,6 +985,20 @@ export class Contract extends Artifact {
 
     this.bytecodeDebug = ralph.buildDebugBytecode(this.bytecode, this.bytecodeDebugPatch)
     this.codeHashDebug = codeHashDebug
+
+    this.decodedMethods = contract.contractCodec.decodeContract(Buffer.from(bytecode, 'hex')).methods
+  }
+
+  publicFunctions(): FunctionSig[] {
+    return this.functions.filter((_, index) => this.decodedMethods[`${index}`].isPublic)
+  }
+
+  usingPreapprovedAssetsFunctions(): FunctionSig[] {
+    return this.functions.filter((_, index) => this.decodedMethods[`${index}`].usePreapprovedAssets)
+  }
+
+  usingAssetsInContractFunctions(): FunctionSig[] {
+    return this.functions.filter((_, index) => this.decodedMethods[`${index}`].useContractAssets)
   }
 
   // TODO: safely parse json
@@ -988,7 +1045,7 @@ export class Contract extends Artifact {
       result.codeHashDebug,
       result.fields,
       result.events,
-      result.functions,
+      result.functions.map(fromFunctionSig),
       result.constants,
       result.enums,
       structs,
@@ -1349,7 +1406,7 @@ export class Script extends Artifact {
       result.bytecodeTemplate,
       result.bytecodeDebugPatch,
       result.fields,
-      result.functions,
+      result.functions.map(fromFunctionSig),
       structs
     )
   }
@@ -1878,7 +1935,9 @@ function genCodeForType(type: string, structs: Struct[]): { bytecode: string; co
   const { immFields, mutFields } = ralph.calcFieldSize(type, true, structs)
   const loadImmFieldByIndex: Method = {
     isPublic: true,
-    assetModifier: 0,
+    usePreapprovedAssets: false,
+    useContractAssets: false,
+    usePayToContractOnly: false,
     argsLength: 1,
     localsLength: 1,
     returnLength: 1,
@@ -1906,7 +1965,9 @@ function genCodeForType(type: string, structs: Struct[]): { bytecode: string; co
   }
   const destroy: Method = {
     isPublic: true,
-    assetModifier: 2,
+    usePreapprovedAssets: false,
+    useContractAssets: true,
+    usePayToContractOnly: false,
     argsLength: 1,
     localsLength: 1,
     returnLength: 0,
@@ -2009,24 +2070,36 @@ export async function testMethod<
 }
 
 export class RalphMap<K extends Val, V extends Val> {
+  private readonly groupIndex: number
   constructor(
     private readonly parentContract: Contract,
-    private readonly parentInstance: ContractInstance,
+    private readonly parentContractId: HexString,
     private readonly mapName: string
-  ) {}
+  ) {
+    this.groupIndex = groupOfAddress(addressFromContractId(parentContractId))
+  }
 
   async get(key: K): Promise<V | undefined> {
-    return getMapItem(this.parentContract, this.parentInstance, this.mapName, key)
+    return getMapItem(this.parentContract, this.parentContractId, this.groupIndex, this.mapName, key)
   }
 
   async contains(key: K): Promise<boolean> {
     return this.get(key).then((v) => v !== undefined)
   }
+
+  toJSON() {
+    return {
+      parentContractId: this.parentContractId,
+      mapName: this.mapName,
+      groupIndex: this.groupIndex
+    }
+  }
 }
 
 export async function getMapItem<R extends Val>(
   parentContract: Contract,
-  parentInstance: ContractInstance,
+  parentContractId: HexString,
+  groupIndex: number,
   mapName: string,
   key: Val
 ): Promise<R | undefined> {
@@ -2036,13 +2109,7 @@ export async function getMapItem<R extends Val>(
     throw new Error(`Map ${mapName} does not exist in contract ${parentContract.name}`)
   }
   const [keyType, valueType] = ralph.parseMapType(mapType)
-  const mapItemContractId = calcWrapperContractId(
-    parentInstance.contractId,
-    index!,
-    key,
-    keyType,
-    parentInstance.groupIndex
-  )
+  const mapItemContractId = calcWrapperContractId(parentContractId, index!, key, keyType, groupIndex)
   const mapItemAddress = addressFromContractId(mapItemContractId)
   try {
     const state = await getCurrentNodeProvider().contracts.getContractsAddressState(mapItemAddress)
